requirements.txt
flask==2.3.3
google-cloud-storage==2.13.0
werkzeug==2.3.7
gunicorn==21.2.0
requests==2.31.0
